+++ src/_info/vendored/mod.rs
@@ -10,6 +10,7 @@
 /// Detailed list of modifications in adapted work.
 //
 // MIT OR Apache-2.0
+crate::doc_! { vendor_mod: "argv", argv }
 crate::doc_! { vendor_mod: "bytehound-preload", bytehound_preload }
 crate::doc_! { vendor_mod: "crossbeam-utils", crossbeam_utils }
 crate::doc_! { vendor_mod: "etcetera", etcetera}