+++ src/sys/env/arg/mod.rs
@@ -2,3 +2,21 @@
 //
 //! Parse arguments to the program.
 //
+
+#[cfg(all(feature = "std", feature = "unsafe_ffi"))]
+#[cfg_attr(nightly_doc, doc(cfg(all(feature = "std", feature = "unsafe_ffi"))))]
+mod os_ref; // ArgsOsRefIter + TEMP args_os_ref_iter (make impl for Env
+
+crate::items! { // structural access: _mods, _all
+    #[allow(unused)]
+    pub use _mods::*;
+
+    mod _mods { #![allow(unused)]
+        #[cfg(all(feature = "std", feature = "unsafe_ffi"))]
+        pub use super::os_ref::*;
+    }
+    pub(super) mod _all {
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+}